requirements.txt
absl-py==1.4.0
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
Django==4.2.3
executing==1.2.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
jax==0.4.13
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mysql==0.0.3
mysqlclient==2.2.0
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.14.1
